Proposing some changes to 2 sections
This all looks great. I added some more text to the "Payment Source" section and the "Batches" section to flesh out a little bit. 

For payment sources, I think it's fair to say that all systems adding donations via the API really should identify themselves as a payment source. (Actually, isn't that required?) 

For Batches, I wanted to flesh out the explanation of what it means for donations to exist in an "uncommitted" batch and what the practical meaning of "visible to donors" means. Hopefully it adds a little more context as to why this concept exists before we dive into the 2 implementation strategies listed immediately thereafter.