fix(BatchVertex): update batch commit doc
It has been brought to our attention by an API user that the docs for
BatchVertex contain an error. These docs describe how to use Batches
but neglect to acknowledge batch policy validation that states a batch
must contain at least one donation in order to be committed. Let's
explicitly add this to the docs to avoid confusion.